pyproject.toml
[tool.poetry]
name = "estour"
version = "0.1.0"
description = ""
authors = ["alexandrejastrow <alexandre.jastrow@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
psycopg2-binary = "^2.9.3"
Pillow = "^9.1.0"
django-autoslug = "^1.9.8"
django-model-utils = "^4.2.0"
django-allauth = "^0.50.0"
django-crispy-forms = "^1.14.0"
django-environ = "^0.8.1"
whitenoise = "^6.0.0"
gunicorn = "^20.1.0"
django-cloudinary-storage = "^0.3.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
